+++ bz-database/attributes/postgres.rb
@@ -1,0 +1,2 @@
+default['bz-database']['postgres']['username'] = 'postgres'
+default['postgresql']['password']['postgres'] = node['bz-database']['postgres']['password']