+++ bz-database/metadata.rb
@@ -4,7 +4,8 @@ maintainer_email 'info@bitzesty.com'
 license          'All rights reserved'
 description      'Database configurationr recipe'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.1.4'
+version          '0.1.5'
 
 depends          'database'
 depends          'mongodb'
+depends          'postgresql'